requirements.txt
flask
flask-cors
python-dotenv
groq
langchain
langchain-community
langchain-huggingface
faiss-cpu
pypdf
